Rename pkg/fjarm/users/v1 to pkg/fjarm/users/usersv1
