+++ api/pkg/fjarm/users/usersv1/BUILD.bazel
@@ -1,12 +1,13 @@
 load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "users",
+    name = "usersv1",
     srcs = [
         "common_errors.go",
         "users.go",
+        "utils.go",
     ],
-    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/users/v1",
+    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/users/usersv1",
     visibility = ["//:__subpackages__"],
     deps = [
         "@build_buf_gen_go_fjarm_fjarm_protocolbuffers_go//fjarm/users/v1:users",
@@ -15,7 +16,7 @@ go_library(
 )
 
 go_test(
-    name = "users_test",
+    name = "usersv1_test",
     srcs = [
         "create_user_request_test.go",
         "user_email_address_test.go",
@@ -25,8 +26,7 @@ go_test(
         "user_password_test.go",
         "user_test.go",
     ],
-    embed = [":users"],
-    importpath = "github.com/fjarm/fjarm/api/pkg/fjarm/users/v1",
+    embed = [":usersv1"],
     deps = [
         "@build_buf_gen_go_fjarm_fjarm_protocolbuffers_go//fjarm/idempotency/v1:idempotency",
         "@build_buf_gen_go_fjarm_fjarm_protocolbuffers_go//fjarm/users/v1:users",
