+++ api/pkg/fjarm/users/usersv1/common_errors.go
@@ -1,4 +1,4 @@
-package v1
+package usersv1
 
 import "fmt"
