+++ api/pkg/fjarm/users/usersv1/create_user_request_test.go
@@ -1,4 +1,4 @@
-package v1
+package usersv1
 
 import (
 	idempotencypb "buf.build/gen/go/fjarm/fjarm/protocolbuffers/go/fjarm/idempotency/v1"
