+++ api/pkg/fjarm/users/usersv1/user_full_name_test.go
@@ -1,4 +1,4 @@
-package v1
+package usersv1
 
 import (
 	userservicepb "buf.build/gen/go/fjarm/fjarm/protocolbuffers/go/fjarm/users/v1"
