+++ api/pkg/fjarm/users/usersv1/user_id_test.go
@@ -1,4 +1,4 @@
-package v1
+package usersv1
 
 import (
 	pb "buf.build/gen/go/fjarm/fjarm/protocolbuffers/go/fjarm/users/v1"
